+++ .vscode/settings.json
@@ -1,8 +1,3 @@
 {
-    "cSpell.words": [
-        "behaviour",
-        "geoapi",
-        "pathify",
-        "prefetch"
-    ]
+    "cSpell.words": ["behaviour", "geoapi", "pathify", "prefetch", "vetur"]
 }
