+++ packages/ramp-core/src/api/index.ts
@@ -1,0 +1,5 @@
+import Map from './map';
+
+export default {
+    Map
+};
