+++ packages/ramp-core/src/store/modules/fixture/fixture-state.ts
@@ -12,21 +12,21 @@ export interface Fixture {
     id: string;
 
     /**
-     * [Optional] Called synchronously when the feature-block is added to R4MP.
+     * [Optional] Called synchronously when the fixture is added to R4MP.
      *
      * @memberof Fixture
      */
     added?(): void;
 
     /**
-     * [Optional] Called synchronously when the feature-block is removed from R4MP.
+     * [Optional] Called synchronously when the fixture is removed from R4MP.
      *
      * @memberof Fixture
      */
     removed?(): void;
 
     /**
-     * [Optional] Called after the core map logic has executed. If the feature-block is added after the map initialization, this is called immediately.
+     * [Optional] Called after the core map logic has executed. If the fixture is added after the map initialization, this is called immediately.
      *
      * At this point, any custom content (panels, on-map components, etc.) can be added to R4MP. Custom content can also be added throughout the life of the [module].
      *
@@ -35,7 +35,7 @@ export interface Fixture {
     initialized?(): void;
 
     /**
-     * [Optional] Called before the feature-block is removed from R4MP.
+     * [Optional] Called before the fixture is removed from R4MP.
      *
      * This is a clean-up phase, and at this point, any custom content (panels, on-map components, etc.) must be removed.
      *
