+++ packages/ramp-core/src/store/modules/fixture/index.ts
@@ -1,0 +1,2 @@
+export * from './fixture-state';
+export * from './fixture-store';
