--- packages/ramp-core/src/store/modules/fixtures/index.ts
@@ -1,2 +1,0 @@
-export * from './fixtures-state';
-export * from './fixtures';
